Refactor OgMembershipStateCacheContextTest so parts of it can be reused for other tests.
